requirements.txt
beautifulsoup4==4.10.0
certifi==2020.6.20
charset-normalizer==2.0.9
html5lib==1.1
idna==3.3
pip==21.2.4
PyYAML==6.0
requests==2.26.0
setuptools==58.0.4
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.7
webencodings==0.5.1
wheel==0.37.0
